Ready to pull into markgar version
Removed github actions test, there seems to be a bug in the existing nuget build samples (link will be added to PR)
Reverted CIS4Azure.Tests.Pester.nuspec to be more similar to existing one in mark repo. Updated schema version, added license, etc.
This should be the initial version of the azure pipeline spec. Instructions on how to add to AzDO will be added to PR.